pyproject.toml
[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.1",
    "coverage>=7.2.5",
]
[project]
name = "struct-gpt"
version = "2.2.0"
description = "structured llm outputs"
authors = [
    {name = "Stephan Fitzpatrick", email = "knowsuchagency@gmail.com"},
]
dependencies = [
    "openai>=0.27.6",
    "pydantic>=1.10.7",
    "ruamel-yaml>=0.17.24",
]
requires-python = ">=3.10"
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Source = "https://github.com/knowsuchagency/struct-gpt"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
